--- errors/index.js
@@ -1,17 +1,0 @@
-const Http403 = require("./Http403");
-const Http404 = require("./Http404");
-const Http405 = require("./Http405");
-const Http422 = require("./Http422");
-const Http500 = require("./Http500");
-
-/**
- * Exports
- * @type {Object} exports
- */
-module.exports = {
-  Http403,
-  Http404, 
-  Http405,
-  Http422,
-  Http500
-};
