+++ invoke.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-./node_modules/.bin/lambda-local -l "./index.js" -h handler -e "./$1/$2/event.json"
+./node_modules/.bin/lambda-local -l "./index.js" -h handler -e "./paths/$1/$2/event.json"
