+++ lib/index.js
@@ -1,0 +1,5 @@
+const HttpStatus = require("./HttpStatus");
+
+module.exports = {
+  HttpStatus
+};
