+++ paths/total-cost/get/event.json
