+++ paths/total-cost/put/DynamoDB.js
