+++ responses/Http500.js
@@ -1,0 +1,36 @@
+const Response = require("./Response");
+
+/**
+ * HTTP Status Code
+ * 
+ * @type {String}
+ */
+const status = "500";
+
+/**
+ * Http500
+ * 
+ * @type {class}
+ */
+module.exports = class extends Response {
+  /**
+   * [STATUS]
+   *
+   * @return {String} #status
+   */
+  static get STATUS() {
+    return status;
+  }
+
+  /**
+   * Http500
+   * 
+   * @return {Http500}  instance
+   */
+  constructor() {
+    super();
+
+    this.statusDescription = "Internal Server Error";
+    this.status = status;
+  }
+};
